+++ src/main/java/draylar/tiered/api/TieredItemTags.java
@@ -7,6 +7,7 @@ import net.minecraft.util.Identifier;
 
 public class TieredItemTags {
 
+    public static final TagKey<Item> MODIFIER_RESTRICTED = register("modifier_restricted");
     public static final TagKey<Item> REFORGE_ADDITION = register("reforge_addition");
     public static final TagKey<Item> REFORGE_BASE_ITEM = register("reforge_base_item");
     public static final TagKey<Item> MAIN_OFFHAND_ITEM = register("main_offhand_item");